Add files via upload
If the script is being run via a non-interactive environment (e.g. if you're connected to the client using WinRM), it can't prompt the user for input. This issue was addressed in the latest update.

Hades will now use script arguments to supply these values, instead of prompting the user for input.